Move database modules into common subdirectory
