+++ api.c
@@ -18,7 +18,7 @@
 #include "log.h"
 #include "request.h"
 #include "config.h"
-#include "database.h"
+#include "database/database.h"
 #include "overTime.h"
 #include "api.h"
 #include "version.h"
