+++ request.c
@@ -16,7 +16,7 @@
 #include "socket.h"
 #include "resolve.h"
 #include "regex_r.h"
-#include "networktable.h"
+#include "database/networktable.h"
 #include "log.h"
 
 bool __attribute__((pure)) command(const char *client_message, const char* cmd) {
